requirements.txt
google-adk
google-genai
python-dotenv
pydantic>=2
requests
rapidfuzz
